Modify the 1MHz clock generation for the external LEDs
A clock divider is used, dividing the unity clock (50MHz) by 50. It is
tested, and works correctly.

The output to the board LEDs is moved to the leds_controller as well.

The data output for the pixels is not the desired one. Apparently,
only the last bit of the pixel frame is sent through the wire,
according to the oscilloscope.